Remove inaccurate translation of "robot"
In Czech, the word "robot" always refers to a machine.  The only time
I can think of when it is used to describe people, it is used to liken
them to a machine - much like it is common in English to use phrases
such as "robotic voice" or "[they] move like robots".  Such comparisons
(in Czech or English) in no way imply slavery.